requirements.txt
Django==5.2.1
django-guest-user==0.5.5
python-dotenv==1.1.0
